+++ CHANGELOG.md
@@ -24,6 +24,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The skip to content link is now underlined.
 - Service support users can now edit all contacts.
 - URN and UKPRN fields no longer autocomplete.
+- When creating a new project and submitting an empty form, the errors are now
+  shown correctly.
 
 ## [Release-70][Release-70]